warp-pipe-core/src/main/java/org/luizricardo/warppipe/WarpFilter.java
package org.luizricardo.warppipe;

import javax.servlet.Filter;
import javax.servlet.FilterChain;
import javax.servlet.FilterConfig;
import javax.servlet.ServletException;
import javax.servlet.ServletRequest;
import javax.servlet.ServletResponse;
import java.io.IOException;
import java.nio.charset.Charset;
import java.nio.charset.StandardCharsets;

public class WarpFilter implements Filter {

    private Charset encoding = StandardCharsets.UTF_8;

    @Override
    public void init(final FilterConfig filterConfig) throws ServletException {
        final String encodingParam = filterConfig.getInitParameter("encoding");
        if (encodingParam != null && !encodingParam.isEmpty()) {
            this.encoding = Charset.forName(encodingParam);
        }
    }

    @Override
    public void doFilter(final ServletRequest request, final ServletResponse response,
                         final FilterChain chain) throws IOException, ServletException {
        chain.doFilter(request, response);
    }

    @Override
    public void destroy() {
    }
}
